fix: fail ingestion when concurrent tasks fail
